pipeline/requirements.txt
python-chess==1.999
stockfish==3.28.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
click==8.1.8
tqdm==4.67.1
nanoid==2.0.0
